moved, renamed, and refactored a bit the search file, formerly the app.js file, to utils folder, in routes folder
